clean RDF declaration using MolData3D
