+++ Code/GraphMol/Descriptors/MolData3Ddescriptors.cpp
@@ -3,6 +3,10 @@
 #include "MolData3Ddescriptors.h"
 #include <GraphMol/RDKitBase.h>
 
+#include "GraphMol/PartialCharges/GasteigerCharges.h"
+#include "GraphMol/PartialCharges/GasteigerParams.h"
+
+
 
 using namespace std;
 
@@ -94,6 +98,13 @@ std::vector<double> MolData3Ddescriptors::GetRelativeIonPol(const RDKit::ROMol& 
 }
 
 
+std::vector<double> MolData3Ddescriptors::GetCharges(const RDKit::ROMol &mol) {
+  std::vector<double> charges(mol.getNumAtoms(), 0);
+  // use 12 iterations... can be more
+  RDKit::computeGasteigerCharges(mol, charges, 12, true);
+  return charges;
+}
+
 int MolData3Ddescriptors::GetPrincipalQuantumNumber(int AtomicNum) {
   if (AtomicNum<=2) return 1;
   else if (AtomicNum<=10) return 2;
@@ -126,4 +137,84 @@ std::vector<double> MolData3Ddescriptors::GetIState(const  RDKit::ROMol &mol){
 
  return Is;
 }
+
+
+// adaptation from EState.py 
+// we need the Is value only there
+std::vector<double> MolData3Ddescriptors::GetEState(const  RDKit::ROMol &mol){
+  int numAtoms = mol.getNumAtoms();
+  std::vector<double> Is;
+
+  for (int i = 0; i < numAtoms; ++i) {
+    const RDKit::Atom * atom= mol.getAtomWithIdx(i);
+    int atNum=atom->getAtomicNum();
+    int d = atom->getDegree();
+    if (d>0 and atNum>1) {
+      int h = atom->getTotalNumHs();
+      int dv = RDKit::PeriodicTable::getTable()->getNouterElecs(atNum)-h;
+      int N = GetPrincipalQuantumNumber(atNum); // period value
+      Is.push_back(round(1000*(4.0/(N*N)*dv+1.0)/d)/1000);  // WHIM-P5.pdf paper 1997  => +7 & NoHydrogens is used!
+    }
+    else Is.push_back(1.0); // Is it zero or one for H!
+ }
+
+  double tmp,p;
+  double *dist =  RDKit::MolOps::getDistanceMat(mol,false,false);
+  double accum[numAtoms];
+  for (int i=0;i<numAtoms;i++) {
+    for (int j=i+1;j<numAtoms;j++) {
+       p = dist[i * numAtoms + j]+1;
+       if (p < 1e6) {
+        tmp = (Is[i] - Is[j]) / (p * p);
+        accum[i] += tmp;
+        accum[j] -= tmp;
+      }
+    }
+  }
+
+ for (int i=0;i<numAtoms;i++) {
+    Is[i]+=accum[i];
+ }
+
+ return Is;
+}
+
+// modification of previous code to follow documentation from Padel code
+std::vector<double> MolData3Ddescriptors::GetEState2(const  RDKit::ROMol &mol){
+  int numAtoms = mol.getNumAtoms();
+  std::vector<double> Is;
+  double Cistate=0.0;
+  int numC=0;
+  double tmp;
+  for (int i = 0; i < numAtoms; ++i) {
+    const RDKit::Atom * atom= mol.getAtomWithIdx(i);
+    int atNum=atom->getAtomicNum();
+    int d = atom->getDegree();
+
+    if (d>0 and atNum>1) {
+      int h = atom->getTotalNumHs();
+      int Zv= RDKit::PeriodicTable::getTable()->getNouterElecs(atNum);
+      double deltaval = (double) (Zv-h);
+      //deltaval = deltaval / (double) (atNum-Zv-1);
+      int N = GetPrincipalQuantumNumber(atNum); // period value
+      tmp=round(1000.0*(4.0/(N*N)*deltaval+1.0)/d)/1000.0;
+      Is.push_back(tmp); 
+      // normalize factore determination
+       if (atNum==6) {
+        Cistate+=tmp;
+        numC++;
+       }
+    }
+    else Is.push_back(1.0);
+ }
+
+
+ double meanCistate= Cistate/numC;
+ for (int i=0;i<numAtoms;i++) {
+    Is[i]=Is[i]/meanCistate;
+ }
+
+ return Is;
+}
+
