+++ Code/GraphMol/Descriptors/MolData3Ddescriptors.h
@@ -17,6 +17,7 @@ class MolData3Ddescriptors
 
 	public:
 	   MolData3Ddescriptors();
+	   std::vector<double> GetCharges(const RDKit::ROMol& mol);
 	   std::vector<double> GetRelativeMW(const RDKit::ROMol& mol);
 	   std::vector<double> GetRelativePol(const RDKit::ROMol& mol);
 	   std::vector<double> GetRelativeRcov(const RDKit::ROMol& mol);
@@ -26,6 +27,8 @@ class MolData3Ddescriptors
 	   std::vector<double> GetUn(int numAtoms);
 	   int GetPrincipalQuantumNumber(int AtomicNum);
 	   std::vector<double> GetIState(const  RDKit::ROMol& mol);
+	   std::vector<double> GetEState(const  RDKit::ROMol& mol);
+	   std::vector<double> GetEState2(const  RDKit::ROMol& mol);
 
 
 };
