+++ Code/GraphMol/Descriptors/RDF.cpp
@@ -55,9 +55,7 @@
 #include <GraphMol/MolTransforms/MolTransforms.h>
 
 #include "RDF.h"
-
-#include "GraphMol/PartialCharges/GasteigerCharges.h"
-#include "GraphMol/PartialCharges/GasteigerParams.h"
+#include "MolData3Ddescriptors.h"
 
 #include <Numerics/Matrix.h>
 #include <Numerics/SquareMatrix.h>
@@ -68,47 +66,14 @@
 
 // data checked using book Todeschini R., Consonni V. - Molecular Descriptors for Chemoinformatics 2009 atomic properties page 21/22
 
-double Pol[] = {
-    0.67,  0,    24.3, 5.60,  3.03, 1.76,  1.10, 0.80, 0.56, 0,    23.6, 10.6,
-    6.80,  5.38, 3.63, 2.90,  2.18, 0,     43.4, 22.8, 0,    0,    0,    11.60,
-    9.40,  8.40, 7.50, 6.80,  6.10, 7.10,  8.12, 6.07, 4.31, 3.77, 3.05, 0,
-    47.3,  27.6, 0,    0,     0,    12.80, 0,    0,    0,    0,    7.20, 7.20,
-    10.20, 7.70, 6.60, 5.50,  5.35, 0,     0,    0,    0,    0,    0,    0,
-    0,     0,    0,    23.50, 0,    0,     0,    0,    0,    0,    0,    0,
-    0,     0,    0,    0,     0,    6.50,  5.80, 5.70, 7.60, 6.80, 7.40};
-double ElectroNeg[] = {
-    2.59, 0,    0.89, 1.81, 2.28, 2.75, 3.19, 3.65, 4.0,  0,    0.56, 1.32,
-    1.71, 2.14, 2.52, 2.96, 3.48, 0,    0.45, 0.95, 0,    0,    0,    1.66,
-    2.2,  2.2,  2.56, 1.94, 1.98, 2.23, 2.42, 2.62, 2.82, 3.01, 3.22, 0,
-    0.31, 0.72, 0,    0,    0,    1.15, 0,    0,    0,    0,    1.83, 1.98,
-    2.14, 2.3,  2.46, 2.62, 2.78, 0,    0,    0,    0,    0,    0,    0,
-    0,    0,    0,    2.0,  0,    0,    0,    0,    0,    0,    0,    0,
-    0,    0,    0,    0,    0,    2.28, 2.54, 2.2,  2.25, 2.29, 2.34};
-
-double VdW[] = {
-    6.71,  0,     25.25, 0.0,   17.88, 22.45, 15.6,  11.49, 9.20,   0,     49.0,
-    21.69, 36.51, 31.98, 26.52, 24.43, 22.45, 0,     87.11, 0.0,   0,     0,
-    0,     44.6,  43.4,  41.05, 35.04, 17.16, 11.49, 11.25, 27.39, 28.73, 26.52,
-    28.73, 31.06, 0,     0.0,   0.0,   0,     0,     0,     33.51, 0,     0,
-    0,     0,     21.31, 16.52, 30.11, 45.83, 38.79, 36.62, 32.52, 0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     72.78, 0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    22.45, 19.16, 15.6,  31.54, 34.53, 38.79};
-
-double  IonPolarizability[] = {
-  13.598, 0, 5.392, 9.323, 8.298, 11.26, 14.534, 13.618, 17.423, 0, 5.139, 7.646,
-  5.986, 8.152, 10.487, 10.360, 12.968, 0, 4.341, 6.113, 0, 0, 0, 6.767, 7.434,
-  7.902, 7.881, 7.640, 7.723, 9.394, 5.999, 7.900, 9.815, 9.752, 11.814, 0, 4.177,
-  5.695, 0, 0, 0, 7.092, 0, 0, 0, 0, 7.576, 8.994, 5.786, 7.344, 8.640, 9.010,
-  10.451, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6.15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-  0, 0, 0, 9.000, 9.226, 10.438, 6.108, 7.417, 7.289};
-
-
 
 
 namespace RDKit {
 namespace Descriptors {
 namespace {
+
+  MolData3Ddescriptors moldata3D;
+
 
 std::vector<double> getG(int n) {
   std::vector<double> res(n);
@@ -118,191 +83,7 @@ std::vector<double> getG(int n) {
   return res;
 }
 
-
-// adaptation from EState.py 
-// check with Molecular Descriptors for Chemoinformatics, Volume 41 (2 Volume Set) page 283/284
-int GetPrincipalQuantumNumber(int AtomicNum) {
-  if (AtomicNum<=2) return 1;
-  else if (AtomicNum<=10) return 2;
-  else if (AtomicNum<=18) return 3;
-  else if (AtomicNum<=36) return 4;
-  else if (AtomicNum<=54) return 5;
-  else if (AtomicNum<=86) return 6;
-  else return 7;
-}
-
-// adaptation from EState.py 
-// we need the Is value only there
-std::vector<double> GetIState(const ROMol &mol){
-  int numAtoms = mol.getNumAtoms();
-  std::vector<double> Is;
-  double Cistate=0.0;
-  int numC=0;
-  double tmp;
-  for (int i = 0; i < numAtoms; ++i) {
-    const RDKit::Atom * atom= mol.getAtomWithIdx(i);
-    int atNum=atom->getAtomicNum();
-    int d = atom->getDegree();
-
-    if (d>0 and atNum>1) {
-      int h = atom->getTotalNumHs();
-      int Zv= RDKit::PeriodicTable::getTable()->getNouterElecs(atNum);
-      double deltaval = (double) (Zv-h);
-      //deltaval = deltaval / (double) (atNum-Zv-1);
-      int N = GetPrincipalQuantumNumber(atNum); // period value
-      tmp=round(1000.0*(4.0/(N*N)*deltaval+1.0)/d)/1000.0;
-      Is.push_back(tmp);  // WHIM-P5.pdf paper 1997  => +7 & NoHydrogens is used!
-      // normalize factore determination
-       if (atNum==6) {
-        Cistate+=tmp;
-        numC++;
-       }
-    }
-    else Is.push_back(1.0);
- }
-
-
-double meanCistate= Cistate/numC;
- for (int i=0;i<numAtoms;i++) {
-    Is[i]=Is[i];
- }
-
- return Is;
-}
-
-// adaptation from EState.py 
-// we need the Is value only there
-std::vector<double> GetIStateRDKit(const ROMol &mol){
-  int numAtoms = mol.getNumAtoms();
-  std::vector<double> Is;
-
-  for (int i = 0; i < numAtoms; ++i) {
-    const RDKit::Atom * atom= mol.getAtomWithIdx(i);
-    int atNum=atom->getAtomicNum();
-    int d = atom->getDegree();
-    if (d>0 and atNum>1) {
-      int h = atom->getTotalNumHs();
-      int dv = RDKit::PeriodicTable::getTable()->getNouterElecs(atNum)-h;
-      int N = GetPrincipalQuantumNumber(atNum); // period value
-      Is.push_back(round(1000*(4.0/(N*N)*dv+1.0)/d)/1000);  // WHIM-P5.pdf paper 1997  => +7 & NoHydrogens is used!
-    }
-    else Is.push_back(1.0);
- }
- // debug mode!
- /*
-for (int i = 0; i < numAtoms; ++i) {
-
-  std::cout << Is[i] << "-" << mol.getAtomWithIdx(i)->getSymbol() << ",";
-}*/
-  double tmp,p;
-  double *dist = MolOps::getDistanceMat(mol,false,false);
-  double accum[numAtoms];
-  for (int i=0;i<numAtoms;i++) {
-    for (int j=i+1;j<numAtoms;j++) {
-       p = dist[i * numAtoms + j]+1;
-       if (p < 1e6) {
-        tmp = (Is[i] - Is[j]) / (p * p);
-        accum[i] += tmp;
-        accum[j] -= tmp;
-      }
-    }
-  }
-
- for (int i=0;i<numAtoms;i++) {
-    Is[i]+=accum[i];
- }
-
- return Is;
-}
-
-std::vector<double> GetCharges(const ROMol &mol) {
-  std::vector<double> charges(mol.getNumAtoms(), 0);
-  // use 12 iterations... can be more
-  computeGasteigerCharges(mol, charges, 12, true);
-  return charges;
-}
-
-std::vector<double> GetRelativePol(const ROMol &mol) {
-  int numAtoms = mol.getNumAtoms();
-
-  std::vector<double> pol(numAtoms, 0);
-  for (int i = 0; i < numAtoms; ++i) {
-    pol[i] = Pol[mol.getAtomWithIdx(i)->getAtomicNum()-1] / Pol[5];
-  }
-
-  return pol;
-}
-
-std::vector<double> GetRelativeIonPol(const ROMol &mol) {
-  int numAtoms = mol.getNumAtoms();
-
-  std::vector<double> ionpol(numAtoms, 0);
-  for (int i = 0; i < numAtoms; ++i) {
-    ionpol[i] = IonPolarizability[mol.getAtomWithIdx(i)->getAtomicNum()-1] / IonPolarizability[5];
-  }
-
-  return ionpol;
-}
-
-
-std::vector<double> GetRelativeElectroNeg(const ROMol &mol) {
-  int numAtoms = mol.getNumAtoms();
-
-  std::vector<double> REN(numAtoms, 0);
-  for (int i = 0; i < numAtoms; ++i) {
-    REN[i] = ElectroNeg[mol.getAtomWithIdx(i)->getAtomicNum()-1] / ElectroNeg[5];
-  }
-
-  return REN;
-}
-
-std::vector<double> GetRelativeVdW(const ROMol &mol) {
-  int numAtoms = mol.getNumAtoms();
-
-  std::vector<double> vdw(numAtoms, 0);
-  for (int i = 0; i < numAtoms; ++i) {
-    vdw[i] = VdW[mol.getAtomWithIdx(i)->getAtomicNum()-1] / VdW[5];
-  }
-
-  return vdw;
-}
-
-std::vector<double> GetAbsPol(const ROMol &mol) {
-  int numAtoms = mol.getNumAtoms();
-  std::vector<double> pol(numAtoms, 0);
-  for (int i = 0; i < numAtoms; ++i) {
-    pol[i] = Pol[mol.getAtomWithIdx(i)->getAtomicNum()-1];
-  }
-
-  return pol;
-}
-
-
-std::vector<double> CalcUnweightedRDF(const ROMol &mol, const Conformer &conf) {
-  int numAtoms = conf.getNumAtoms();
-  int confId = conf.getId();
-
-  std::vector<double> R = getG(30);
-
-  std::vector<double> RDFres;
-  //std::vector<std::vector<double> > DM = GetGeometricalDistanceMatrix(points);
-  double *DM = MolOps::get3DDistanceMat(mol,confId);
-
-  for (int i = 0; i < 30; i++) {
-    double res = 0;
-    for (int j = 0; j < numAtoms - 1; j++) {
-      for (int k = j + 1; k < numAtoms; k++) {
-        res += exp(-100 * pow(R[i] - DM[j * numAtoms + k], 2));
-          
-      }
-    }
-
-    RDFres.push_back(round( 1000 * res) / 1000);
-  }
-
-  return RDFres;
-}
-
+/*
 std::vector<double> CalcChargeRDF(const ROMol &mol, const Conformer &conf) {
   int numAtoms = conf.getNumAtoms();
   int confId = conf.getId();
@@ -326,7 +107,36 @@ std::vector<double> CalcChargeRDF(const ROMol &mol, const Conformer &conf) {
   }
 
   return RDFres;
+
+*/
+
+
+std::vector<double> CalcUnweightedRDF(const ROMol &mol, const Conformer &conf) {
+  int numAtoms = conf.getNumAtoms();
+  int confId = conf.getId();
+
+  std::vector<double> R = getG(30);
+
+  std::vector<double> RDFres;
+
+  double *DM = MolOps::get3DDistanceMat(mol,confId);
+
+  for (int i = 0; i < 30; i++) {
+    double res = 0;
+    for (int j = 0; j < numAtoms - 1; j++) {
+      for (int k = j + 1; k < numAtoms; k++) {
+        res += exp(-100 * pow(R[i] - DM[j * numAtoms + k], 2));
+          
+      }
+    }
+
+    RDFres.push_back(round( 1000 * res) / 1000);
+  }
+
+  return RDFres;
 }
+
+
 
 std::vector<double> CalcMassRDF(const ROMol &mol, const Conformer &conf) {
   int numAtoms = conf.getNumAtoms();
@@ -334,14 +144,13 @@ std::vector<double> CalcMassRDF(const ROMol &mol, const Conformer &conf) {
 
   std::vector<double> R = getG(30);
   std::vector<double> RDFres;
-  //std::vector<std::vector<double> > DM = GetGeometricalDistanceMatrix(points);
+
   double *DM = MolOps::get3DDistanceMat(mol,confId);
 
 
-  std::vector<double> Mass;
-  for (int p = 0; p < numAtoms; p++) {
-      Mass.push_back(mol.getAtomWithIdx(p)->getMass());
-  }
+  std::vector<double> Mass = moldata3D.GetRelativeMW(mol);
+
+
 
   for (int i = 0; i < 30; i++) {
     double res = 0;
@@ -350,8 +159,7 @@ std::vector<double> CalcMassRDF(const ROMol &mol, const Conformer &conf) {
         res += Mass[j] * Mass[k] * exp(-100 * pow(R[i] - DM[j * numAtoms + k], 2));
       }
     }
-    // center mass to carbon =>  / 12.011 / 12.011
-      RDFres.push_back(round( 1000 * res / 12.011 / 12.011 ) / 1000);
+      RDFres.push_back(round( 1000 * res) / 1000);
   }
 
   return RDFres;
@@ -363,10 +171,10 @@ std::vector<double> CalcPolRDF(const ROMol &mol, const Conformer &conf) {
 
   std::vector<double> R = getG(30);
   std::vector<double> RDFres;
-  //std::vector<std::vector<double> > DM = GetGeometricalDistanceMatrix(points);
+
   double *DM = MolOps::get3DDistanceMat(mol,confId);
 
-  std::vector<double> RelativePol = GetRelativePol(mol);
+  std::vector<double> RelativePol = moldata3D.GetRelativePol(mol);
 
   for (int i = 0; i < 30; i++) {
     double res = 0;
@@ -389,10 +197,10 @@ std::vector<double> CalcIonPolRDF(const ROMol &mol, const Conformer &conf) {
 
   std::vector<double> R = getG(30);
   std::vector<double> RDFres;
-  //std::vector<std::vector<double> > DM = GetGeometricalDistanceMatrix(points);
+
   double *DM = MolOps::get3DDistanceMat(mol,confId);
 
-  std::vector<double> IonPol = GetRelativeIonPol(mol);
+  std::vector<double> IonPol = moldata3D.GetRelativeIonPol(mol);
 
   for (int i = 0; i < 30; i++) {
     double res = 0;
@@ -422,7 +230,7 @@ std::vector<double> CalcIStateRDF(const ROMol &mol, const Conformer &conf) {
 
   double *DM = MolOps::get3DDistanceMat(mol,confId);
 
-  std::vector<double> IState = GetIState(mol);
+  std::vector<double> IState = moldata3D.GetIState(mol);
   std::cout << "IStates:";
 
   for (int j=0;j<numAtoms;j++) {
@@ -437,7 +245,6 @@ std::vector<double> CalcIStateRDF(const ROMol &mol, const Conformer &conf) {
                exp(-100 * pow(R[i] - DM[j * numAtoms + k], 2));
       }
     }
-
       RDFres.push_back(round( 1000 * res) / 1000);
   }
 
@@ -451,10 +258,10 @@ std::vector<double> CalcElectroNegRDF(const ROMol &mol, const Conformer &conf) {
 
   std::vector<double> R = getG(30);
   std::vector<double> RDFres;
-  //std::vector<std::vector<double> > DM = GetGeometricalDistanceMatrix(points);
+
   double *DM = MolOps::get3DDistanceMat(mol,confId);
 
-  std::vector<double> RelativeElectroNeg = GetRelativeElectroNeg(mol);
+  std::vector<double> RelativeElectroNeg = moldata3D.GetRelativeENeg(mol);
 
   for (int i = 0; i < 30; i++) {
     double res = 0;
@@ -479,7 +286,7 @@ std::vector<double> CalcVdWvolRDF(const ROMol &mol, const Conformer &conf) {
   std::vector<double> RDFres;
   double *DM = MolOps::get3DDistanceMat(mol, confId);
 
-  std::vector<double> RelativeVdW = GetRelativeVdW(mol);
+  std::vector<double> RelativeVdW = moldata3D.GetRelativeVdW(mol);
 
   for (int i = 0; i < 30; i++) {
     double res = 0;
@@ -529,8 +336,6 @@ std::vector<double> RDF(const ROMol &mol, int confId) {
   std::vector<double> res8=CalcIStateRDF(mol,conf);
   res1.insert(res1.end(),res8.begin(), res8.end());
 
-
-  
 
  // std::vector<double> res3=CalcChargeRDF(mol,conf);
  // res1.insert(res1.end(),res3.begin(), res3.end());
